'CMakeLists.txt' (MuGirlStau-00-02-35)
	* reducing memory by using floats in StauCalibrationParameters (ATLASRECTS-2881)
	* Tagging as MuGirlStau-00-02-35

2016-02-19 Jochen Meyer
	* fixing CppCheck defects
	* Tagging as MuGirlStau-00-02-34

+++ Reconstruction/MuonIdentification/MuGirlStau/MuGirlStau/StauCalibration.h
@@ -15,9 +15,9 @@ namespace MuGirlNS
 {
     struct StauCalibrationParameters
     {
-        double timeShift;
-        double error;
-        StauCalibrationParameters(double in_timeShift = 0, double in_error = 0) :
+        float timeShift;
+        float error;
+        StauCalibrationParameters(float in_timeShift = 0, float in_error = 0) :
                 timeShift(in_timeShift), error(in_error)
         {
         }
